+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 0.9.11
 
-2022 Feb 20
+2022 Feb 21
 
 * Fixed waveform context menu (QtGui.QMenu -> QtWidgets.QMenu).
 * Fixed range tools on JLS v2 files (except for export) #131